fix floating-point division ambiguity

+++ source/Engine/Math/Random.cpp
@@ -15,7 +15,7 @@ void Random::SetSeed(Sint32 seed) {
 float Random::Get() {
 	Sint32 nextSeed = (1103515245 * Seed + 12345) & 0x7FFFFFFF;
 	Seed = nextSeed;
-	return ((float)nextSeed) / 0x7FFFFFFF;
+	return ((float)nextSeed) / (float)0x7FFFFFFF;
 }
 
 float Random::Max(float max) {
